# docs/swagger.yaml
definitions:
  model.TopupBalanceRequest:
    properties:
      amount:
        type: integer
      phone_number:
        type: string
    type: object
  model.UserLoginRequest:
    properties:
      password:
        type: string
      user_identity:
        type: string
    type: object
  model.UserRegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: ridho otto junior
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ottojunior
  version: "1.0"
paths:
  /account-info:
    get:
      consumes:
      - application/json
      description: get account information from logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get accountt information from logged in user
      tags:
      - emoney-service
  /balance:
    get:
      consumes:
      - application/json
      description: get latest balance from logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get latest balance from logged in user
      tags:
      - emoney-service
  /login:
    post:
      consumes:
      - application/json
      description: login
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: logged in existing user
      tags:
      - emoney-service
  /product:
    get:
      consumes:
      - application/json
      description: get all available product to buy
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get all available product to buy
      tags:
      - emoney-service
  /register:
    post:
      consumes:
      - application/json
      description: register new user
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: register new user
      tags:
      - emoney-service
  /topup:
    post:
      consumes:
      - application/json
      description: handle new topup balance
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TopupBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: handle new topup balance
      tags:
      - topup-service
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"